# Create shading patterns the size of the current path. (bug 1722807)