@brendandahl 2021-11-03
Some notes from performance testing:
1) bug1722807_page2 is MUCH faster going from 62,066ms to 330ms in Firefox and the several GB memory increase is gone
2) bug1721949 is a big slower, but a pretty small difference

-- Grouped By pdf, stat, browser --
pdf                | stat         | browser | Count | Baseline(ms) | Current(ms) |    +/- |     %  | Result(P<.05)
------------------ | ------------ | ------- | ----- | ------------ | ----------- | ------ | ------ | -------------
bug1721949         | Overall      | chrome  |    50 |          136 |         168 |     32 |  23.27 |        slower
bug1721949         | Overall      | firefox |    50 |          175 |         185 |     10 |   5.87 |              
bug1721949         | Page Request | chrome  |    50 |            2 |           2 |      0 |   4.82 |              
bug1721949         | Page Request | firefox |    50 |            2 |           2 |      0 |  -8.57 |              
bug1721949         | Rendering    | chrome  |    50 |          135 |         166 |     32 |  23.56 |        slower
bug1721949         | Rendering    | firefox |    50 |          172 |         183 |     10 |   6.06 |              
bug1722807_page2   | Overall      | chrome  |     5 |          868 |         308 |   -561 | -64.56 |              
bug1722807_page2   | Overall      | firefox |     5 |        62069 |         332 | -61737 | -99.47 |        faster
bug1722807_page2   | Page Request | chrome  |     5 |            6 |           4 |     -2 | -27.59 |              
bug1722807_page2   | Page Request | firefox |     5 |            3 |           2 |     -1 | -38.46 |              
bug1722807_page2   | Rendering    | chrome  |     5 |          863 |         303 |   -559 | -64.83 |              
bug1722807_page2   | Rendering    | firefox |     5 |        62066 |         330 | -61736 | -99.47 |        faster
issue13325_reduced | Overall      | chrome  |    50 |           16 |          16 |      0 |   0.00 |              
issue13325_reduced | Overall      | firefox |    50 |           24 |          11 |    -13 | -54.85 |        faster
issue13325_reduced | Page Request | chrome  |    50 |            1 |           1 |      0 |   9.26 |              
issue13325_reduced | Page Request | firefox |    50 |            1 |           1 |      0 | -37.04 |              
issue13325_reduced | Rendering    | chrome  |    50 |           15 |          15 |      0 |  -0.81 |              
issue13325_reduced | Rendering    | firefox |    50 |           23 |          10 |    -13 | -55.87 |        faster