@Snuffleupagus 2021-11-04
> Are these chrome annotation and form failures new?

I cannot recall having seen them before, unfortunately.