@Snuffleupagus 2021-11-04
> Are these chrome annotation and form failures new?

I cannot recall having seen them before, unfortunately.

> They seem to work fine locally in tests and in the viewer....

When testing this patch in Google Chrome locally, I can reproduce the test-failures. However, there's no errors in the console that'd explain the rendering failures.