@brendandahl 2021-11-05
I haven't fixed the annotation/form bug yet. Just pushing so I can try this out on linux/windows...